Moved sources to app folder (#13)
* Stage (#11)

* Initial scaffolding for site.

* Slight adjustment to Readme content

* Use case and beacon integration (#2)

* Implementation of the demo use-case and adopted Gravity Beacon.

* WIP: Upgraded bootstrap to v4, crude viewer of CV document and improved styling of demo pages.

* Renamed '/application' endpoint to '/inspect'. Styled the content of cv 'employment' and 'qualifications'.

* job application demo form populate with CV data

* Added a cv data input form and som refactoring

* Fixed UX, added save CV button

* Fixed UX around phone field

* updated readme for demosite installation

* Adding dockerfile for easier deployment (#3)

* Added extra EJS property checks.

* Created Dockerfile that starts Nginx and Nodejs. Supervisord handles the processes.

* Updated readme, and removed a comment from Dockerfile

* Updated readme

* Fixed deploy (#6)

* Fixed deploy

* Added possibilities to start without SSL configuration.
And created a .env-example file for convenience.

* removed hostname in server config

* Update README.md

* Separation of concerns (#7)

* Fixed deploy

* Removed all functionality that should not reside in this service/project.

* Update package.json

* Updated .env-example

* Feature/better form (#8)

* Improvements to the cv form

* End-to-end success and design

* Fix/portability url (#9)

* Improvements to the cv form

* End-to-end success and design

* Portability url from environment vars

* Fix logotype. (#10)

* Using envelope and api-key for portability

* Moved sources to app folder
