+++ app/views/pages/index.ejs
@@ -43,7 +43,7 @@
                                 <div id="af-connect-button" class="af-connect-module"
                                     data-af-portability-url="<%= portabilityUrl %>"
                                     data-af-connect-url="https://connectstage.arbetsformedlingen.se/"
-                                    data-api-key="INSERT_API_KEY_HERE"
+                                    data-af-portability-api-key="dummydummydummydummydummydummydummydummydummydummy"
                                     data-on-response="onResponse">
                                 </div>
                             </div>
