+++ package.json
@@ -14,8 +14,8 @@
     "nodemon": "^1.19.2"
   },
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon",
+    "start": "node app/index.js",
+    "dev": "nodemon --watch app app/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": "",
